requirements.txt
confluent_kafka==2.11.1
python-dotenv==1.1.1
